code cleanUp
