+++ src/main/java/com/easyrest/pages/AdminPanel.java
@@ -19,18 +19,6 @@ public class AdminPanel extends BasePage {
     private final By usersName = By.cssSelector("th:nth-child(1)");
     private final By addModeratorButton = By.xpath("//span[text()='Add moderator']");
 
-    /**
-     *  "CREATE MODERATOR ACCOUNT" FORM LOCATORS
-     */
-    private final By nameField = By.xpath("//input[@name='name']");
-    private final By emailField = By.xpath("//input[@name='email']");
-    private final By phoneNumberField = By.xpath("//input[@name='phoneNumber']");
-    private final By birthDateField = By.xpath("//input[@name='birthDate']");
-    private final By passwordField = By.xpath("//input[@name='password']");
-    private final By confirmPasswordField = By.xpath("//input[@name='repeated_password']");
-    private final By createAccountButton = By.xpath("//span[text()='Create account']");
-    private final By cancelButton = By.xpath("//span[text()='Cancel']");
-
     public AdminPanel(WebDriver driver) {
         super(driver);
     }
@@ -47,9 +35,7 @@ public class AdminPanel extends BasePage {
         driver.findElement(activeButton).click();
     }
 
-    public void clickBannedButton(){
-        driver.findElement(bannedButton).click();
-    }
+    public void clickBannedButton(){ driver.findElement(bannedButton).click(); }
 
     public String getFirstNameFromClient() {
         List<WebElement> clientList = driver.findElements(usersName);
@@ -57,48 +43,45 @@ public class AdminPanel extends BasePage {
         return firstNameFromList.getText();
     }
 
-    public AdminPanel clickToMakeFirstClientActive(){
-        clickBannedButton();
+    private void makeActionOnFirstClient() {
         List<WebElement> userList = driver.findElements(actionButton);
         if(userList.size() != 0) {
             userList.get(0).click();
         } else {
-            System.out.println("There is no one banned users.");
+            throw new IndexOutOfBoundsException("There is no one user in list.");
         }
+    }
+
+    public AdminPanel clickToMakeFirstClientActive(){
+        clickBannedButton();
+        makeActionOnFirstClient();
         return this;
     }
 
     public AdminPanel clickToBanFirstClient(){
         clickActiveButton();
-        List<WebElement> userList = driver.findElements(actionButton);
-        if(userList.size() != 0) {
-            userList.get(0).click();
-        } else {
-            System.out.println("There is no one active users.");
-        }
+        makeActionOnFirstClient();
         return this;
+    }
+
+    private boolean findUserInListAfterAction(String name){
+        List<WebElement> usersNames = driver.findElements(usersName);
+        for (WebElement item : usersNames) {
+            if(item.getText() == name){
+                return true;
+            }
+        }
+        return false;
     }
 
     public boolean checkIfClientMovedToBannedList(String name) {
         clickBannedButton();
-        List<WebElement> bannedUsersName = driver.findElements(usersName);
-        for (WebElement item : bannedUsersName) {
-            if(item.getText() == name){
-                return true;
-            }
-        }
-        return false;
+        return findUserInListAfterAction(name);
     }
 
     public boolean checkIfClientMovedToActiveList(String name) {
         clickActiveButton();
-        List<WebElement> activeUsersName = driver.findElements(usersName);
-        for (WebElement item : activeUsersName) {
-            if(item.getText() == name){
-                return true;
-            }
-        }
-        return false;
+        return findUserInListAfterAction(name);
     }
 
     /**
@@ -213,43 +196,6 @@ public class AdminPanel extends BasePage {
         clickModerators();
         driver.findElement(addModeratorButton).click();
         return this;
-    }
-
-    /**
-     *  Methods to fill form "CREATE MODERATOR ACCOUNT"
-     */
-    public AdminPanel inputName (String name) {
-        driver.findElement(nameField).sendKeys(name);
-        return this;
-    }
-
-    public AdminPanel inputEmail (String email) {
-        driver.findElement(emailField).sendKeys(email);
-        return this;
-    }
-
-    public AdminPanel inputPhone(String phoneNumber) {
-        driver.findElement(phoneNumberField).sendKeys(phoneNumber);
-        return this;
-    }
-
-    public AdminPanel inputPassword (String password) {
-        driver.findElement(passwordField).sendKeys(password);
-        driver.findElement(confirmPasswordField).sendKeys(password);
-        return this;
-    }
-
-    public Calendar clickOnBirthdayField() {
-        driver.findElement(birthDateField).click();
-        return new Calendar(driver);
-    }
-
-    public void clickCreateAccount() {
-        driver.findElement(createAccountButton).click();
-    }
-
-    public void clickCancelButton() {
-        driver.findElement(cancelButton).click();
     }
 
     /**
